TEST: Simplify the use of the Null backend
