+++ test/module.mk
@@ -15,19 +15,16 @@ TEST_LIBS += test/null_osystem.o \
 	backends/fs/posix/posix-iostream.o \
 	backends/fs/abstract-fs.o \
 	backends/fs/stdiostream.o \
-	backends/modular-backend.o \
-	test/stubs.o
+	backends/modular-backend.o
 endif
 
 ifdef WIN32
 TEST_LIBS += test/null_osystem.o \
-	backends/fs/posix/posix-fs-factory.o \
-	backends/fs/posix/posix-fs.o \
-	backends/fs/posix/posix-iostream.o \
+	backends/fs/windows/windows-fs-factory.o \
+	backends/fs/windows/windows-fs.o \
 	backends/fs/abstract-fs.o \
 	backends/fs/stdiostream.o \
-	backends/modular-backend.o \
-	test/stubs.o
+	backends/modular-backend.o
 endif
 
 TEST_LIBS +=	audio/libaudio.a math/libmath.a common/libcommon.a
@@ -50,7 +47,6 @@ TEST_CXXFLAGS := $(filter-out -Wglobal-constructors,$(CXXFLAGS))
 
 ifdef WIN32
 TEST_LDFLAGS := $(filter-out -mwindows,$(TEST_LDFLAGS))
-TEST_LIBS += backends/fs/windows/windows-fs-factory.o backends/fs/windows/windows-fs.o
 endif
 
 ifdef N64
